pyproject.toml
[tool.poetry]
name = "aws-chatbot-infra"
version = "0.1.0"
description = ""
authors = ["Krzysztof Banasiak <krzysztof.banasiak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.116.1"
constructs = "^10.3.0"
boto3 = "^1.34.9"
aws-cdk-aws-lambda-python-alpha = "^2.117.0a0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
